refactor: extract SpanProcessor to pkg/processor
Move SpanProcessor from cmd/candela-server/processor.go to
pkg/processor/processor.go so both candela-server and candela-local
can share the same span processing pipeline.

Changes:
- pkg/processor/processor.go: New() constructor (was NewSpanProcessor)
- pkg/processor/processor_test.go: moved 3 tests
- cmd/candela-server/main.go: import pkg/processor, rename var to proc
- Delete cmd/candela-server/processor.go and processor_test.go

Zero behavioral changes — pure refactor.
